Split out threadpool from common.c/.h
