+++ src/geometry/functional.c
@@ -8,6 +8,7 @@
 #include "builtin.h"
 #include "veneer.h"
 #include "common.h"
+#include "threadpool.h"
 #include "error.h"
 #include "value.h"
 #include "object.h"
