+++ src/support/CMakeLists.txt
@@ -1,8 +1,9 @@
 target_sources(morpho
     PRIVATE
-        common.c    common.h
-        memory.c    memory.h
-        parse.c     parse.h
-        random.c    random.h
-        resources.c resources.h
+        common.c     common.h
+        memory.c     memory.h
+        parse.c      parse.h
+        random.c     random.h
+        resources.c  resources.h
+        threadpool.c threadpool.h
 )
