+++ src/support/threadpool.h
@@ -1,0 +1,52 @@
+/** @file threadpool.h
+ *  @author T J Atherton
+ *
+ *  @brief Thread pool 
+ */
+
+#ifndef threadpool_h
+#define threadpool_h
+
+#include <pthread.h>
+#include <stdbool.h>
+#include "varray.h"
+
+/* -----------------------------------------
+ * Tasks
+ * ----------------------------------------- */
+
+/** A task is the basic unit of work that is allocated to the thread pool; it comprises a work function to perform the task, and a single argument. */
+ 
+/** A workfn will be called by the threadpool once a thread is available.
+    You must supply all relevant information for both input and output in a single structure passed as an opaque reference. */
+typedef bool (* workfn) (void *arg);
+
+typedef struct {
+    workfn func;    // Function to call that performs the task
+    void *arg;      // Opaque pointer passed as an argument to the work function
+} task;
+
+DECLARE_VARRAY(task, task);
+
+/* -----------------------------------------
+ * Thread pools
+ * ----------------------------------------- */
+
+typedef struct {
+    pthread_mutex_t lock_mutex; /* Lock for access to threadpool structure. */
+    pthread_cond_t work_available_cond; /* Signals that work is available. */
+    pthread_cond_t work_halted_cond; /* Signals when no threads are processing. */
+    int nprocessing; /* Number of threads actively processing work */
+    int nthreads; /* Number of active threads. */
+    bool stop; /* Indicates threads should terminate */
+
+    varray_task queue; /* Queue of tasks lined up */
+} threadpool;
+
+bool threadpool_init(threadpool *pool, int nworkers);
+void threadpool_clear(threadpool *pool);
+bool threadpool_add_task(threadpool *pool, workfn func, void *arg);
+void threadpool_fence(threadpool *pool);
+void threadpool_wait(threadpool *pool);
+
+#endif /* threadpool_h */
